backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn>=0.30.6
pydantic>=2.9.2
python-dotenv>=1.0.1
python-multipart>=0.0.10

# Database
SQLAlchemy>=2.0.35

# LangChain & LLM
langchain>=0.3.1
langchain-community>=0.3.0
langchain-huggingface>=0.0.3
langchain-chroma>=0.1.4
langchain-ollama
sentence-transformers>=3.0.0

# Vector Store
chromadb>=0.5.5

# Document Processing
PyMuPDF>=1.24.0
pypdf>=4.3.1
python-docx

# OCR
pytesseract>=0.3.10
pypdfium2>=4.30.0
Pillow>=10.2.0

# Streaming
sse-starlette>=2.0.0

# Auth
python-jose[cryptography]
bcrypt

# Testing
pytest>=8.3.3

# Build (Electron)
pyinstaller>=6.10.0
